+++ packages/contracts/lib.esm/index.js
@@ -772,7 +772,7 @@ export class BaseContract {
             logger.throwError("events require a provider or a signer with a provider", Logger.errors.UNSUPPORTED_OPERATION, { operation: "once" });
         }
         runningEvent.addListener(listener, once);
-        // Track this running event and its listeners (may already be there; but no hard in updating)
+        // Track this running event and its listeners (may already be there; but no harm in updating)
         this._runningEvents[runningEvent.tag] = runningEvent;
         // If we are not polling the provider, start polling
         if (!this._wrappedEmits[runningEvent.tag]) {