+++ packages/networks/lib.esm/index.js
@@ -60,6 +60,12 @@ function ethDefaultProvider(network) {
             }
             catch (error) { }
         }
+        if (providers.QuickNodeProvider && options.quicknode !== "-") {
+            try {
+                providerList.push(new providers.QuickNodeProvider(network, options.quicknode));
+            }
+            catch (error) { }
+        }
         if (providerList.length === 0) {
             return null;
         }
@@ -140,6 +146,11 @@ const networks = {
         name: "sepolia",
         _defaultProvider: ethDefaultProvider("sepolia")
     },
+    holesky: {
+        chainId: 17000,
+        name: "holesky",
+        _defaultProvider: ethDefaultProvider("holesky")
+    },
     // ETC (See: #351)
     classic: {
         chainId: 61,
@@ -160,7 +171,11 @@ const networks = {
         name: "matic",
         _defaultProvider: ethDefaultProvider("matic")
     },
-    maticmum: { chainId: 80001, name: "maticmum" },
+    maticmum: {
+        chainId: 80001,
+        name: "maticmum",
+        _defaultProvider: ethDefaultProvider("maticmum")
+    },
     optimism: {
         chainId: 10,
         name: "optimism",
@@ -168,9 +183,11 @@ const networks = {
     },
     "optimism-kovan": { chainId: 69, name: "optimism-kovan" },
     "optimism-goerli": { chainId: 420, name: "optimism-goerli" },
+    "optimism-sepolia": { chainId: 11155420, name: "optimism-sepolia" },
     arbitrum: { chainId: 42161, name: "arbitrum" },
     "arbitrum-rinkeby": { chainId: 421611, name: "arbitrum-rinkeby" },
     "arbitrum-goerli": { chainId: 421613, name: "arbitrum-goerli" },
+    "arbitrum-sepolia": { chainId: 421614, name: "arbitrum-sepolia" },
     bnb: { chainId: 56, name: "bnb" },
     bnbt: { chainId: 97, name: "bnbt" },
 };