+++ packages/providers/lib/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "providers/5.7.2";
+export declare const version = "providers/5.8.0";
 //# sourceMappingURL=_version.d.ts.map