+++ packages/providers/lib/ankr-provider.js
@@ -32,8 +32,18 @@ function getHost(name) {
             return "rpc.ankr.com/eth_rinkeby/";
         case "goerli":
             return "rpc.ankr.com/eth_goerli/";
+        case "sepolia":
+            return "rpc.ankr.com/eth_sepolia/";
         case "matic":
             return "rpc.ankr.com/polygon/";
+        case "maticmum":
+            return "rpc.ankr.com/polygon_mumbai/";
+        case "optimism":
+            return "rpc.ankr.com/optimism/";
+        case "optimism-goerli":
+            return "rpc.ankr.com/optimism_testnet/";
+        case "optimism-sepolia":
+            return "rpc.ankr.com/optimism_sepolia/";
         case "arbitrum":
             return "rpc.ankr.com/arbitrum/";
     }