+++ packages/sha2/lib/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "sha2/5.7.0";
+export declare const version = "sha2/5.8.0";
 //# sourceMappingURL=_version.d.ts.map