+++ packages/tests/lib/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "tests/5.7.1";
+export declare const version = "tests/5.8.0";
 //# sourceMappingURL=_version.d.ts.map