+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 This change log is managed by `admin/cmds/update-versions` but may be manually updated.
 
 
+ethers/v5.8.0 (2024-02-23 19:11)
+--------------------------------
+
+  - Added modern networks, updated third-party backend URLs and added QuickNode. ([#3935](https://github.com/ethers-io/ethers.js/issues/3935), [#4010](https://github.com/ethers-io/ethers.js/issues/4010); [f7c813d](https://github.com/ethers-io/ethers.js/commit/f7c813d3130f4790be4381c06ee5e56c8d513c79))
+  - Updated tests for goerli. ([706d3ca](https://github.com/ethers-io/ethers.js/commit/706d3ca27c233385944877eb64c099c74109b6c2))
+
 ethers/v5.7.2 (2022-10-19 04:19)
 --------------------------------